made some formatting and refactoring
